added views/activities and started working on notifications
