+++ app/src/main/java/com/aden/radq/NotificationsActivity.java
@@ -1,0 +1,14 @@
+package com.aden.radq;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class NotificationsActivity extends AppCompatActivity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.notifications_activity);
+
+    }
+}
