+++ app/src/main/java/com/aden/radq/SettingsActivity.java
@@ -1,0 +1,15 @@
+package com.aden.radq;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class SettingsActivity extends AppCompatActivity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.settings_activity);
+
+
+    }
+}
