+++ app/src/main/res/layout/loading_activity.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="#B6E6E8"
+    android:gravity="center_vertical"
+    android:orientation="vertical">
+
+    <ImageView
+        android:id="@+id/imgRADQ"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginTop="0dp"
+        android:contentDescription="@string/app_name"
+        app:srcCompat="@drawable/radq" />
+
+    <ImageView
+        android:id="@+id/imgAppImage"
+        android:layout_width="311dp"
+        android:layout_height="329dp"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginTop="0dp"
+        android:contentDescription="@string/app_logo"
+        app:srcCompat="@drawable/elder" />
+
+    <TextView
+        android:id="@+id/txtAppDescriptionLoadingPage"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal"
+        android:text="@string/app_description"
+        android:textColor="@android:color/white"
+        android:textSize="18sp" />
+
+</LinearLayout>
