+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,20 @@
 <resources>
     <string name="app_name">Robô Autônomo Detector de Quedas</string>
-    <string name="yolo">YOLO</string>
+    <string name="app_description">Robô autôomo detector de quedas</string>
+    <string name="yolo">Botão Temporário (será removido)</string>
     <string name="downloading_necessary_files">Baixando arquivos necessários do RADQ</string>
     <string name="downloading_CFG_File">Arquivo \'cfg\' da identificação visual</string>
     <string name="downloading_WEIGHT_File">Arquivo \'weights\' da identificação visual</string>
+    <string name="define_safe_contact_button">Definir contato de segurança</string>
+    <string name="settings_title">Configurações</string>
+    <string name="safe_contact_title">Contato de segurança</string>
+    <string name="back_button_text">Voltar</string>
+    <string name="alarms_title">Alarmes</string>
+    <string name="alarms_description">Configure ou adicione alarmes de medicamento</string>
+    <string name="set_alarm">Novo Alarme</string>
+    <string name="notifications_title">Notificações</string>
+    <string name="welcome_title">Bem vindo ao RADQ</string>
+    <string name="welcome_description">Seu robô autônomo detector de quedas</string>
+    <string name="camera">Câmera</string>
+    <string name="app_logo">Uma senhora com uma muleta sendo acompanhada pelo RADQ, Robô Autônomo Detector de Quedas</string>
 </resources>
